restructure project to fix cycle imports
