+++ internal/kademlia/routingtable.go
@@ -1,23 +1,19 @@
-package routing
-
-import (
-	. "github.com/viktorfrom/d7024e-kademlia/internal/kademlia"
-)
+package kademlia
 
 // RoutingTable definition
-// keeps a refrence contact of me and an array of buckets
+// keeps a refrence contact of ME and an array of buckets
 type RoutingTable struct {
-	me      Contact
+	ME      Contact
 	buckets [IDLength * 8]*bucket
 }
 
 // NewRoutingTable returns a new instance of a RoutingTable
-func NewRoutingTable(me Contact) *RoutingTable {
+func NewRoutingTable(ME Contact) *RoutingTable {
 	routingTable := &RoutingTable{}
 	for i := 0; i < IDLength*8; i++ {
 		routingTable.buckets[i] = newBucket()
 	}
-	routingTable.me = me
+	routingTable.ME = ME
 	return routingTable
 }
 
@@ -58,7 +54,7 @@ func (routingTable *RoutingTable) FindClosestContacts(target *KademliaID, count 
 
 // getBucketIndex get the correct Bucket index for the KademliaID
 func (routingTable *RoutingTable) getBucketIndex(id *KademliaID) int {
-	distance := id.CalcDistance(routingTable.me.ID)
+	distance := id.CalcDistance(routingTable.ME.ID)
 	for i := 0; i < IDLength; i++ {
 		for j := 0; j < 8; j++ {
 			if (distance[i]>>uint8(7-j))&0x1 != 0 {
