+++ internal/kademlia/routingtable_test.go
@@ -1,10 +1,8 @@
-package routing
+package kademlia
 
 import (
 	"fmt"
 	"testing"
-
-	. "github.com/viktorfrom/d7024e-kademlia/internal/kademlia"
 )
 
 func TestRoutingTable(t *testing.T) {
